refactor mac specific code
